Getting tracks code added
Code in place to request and parse tracks.  For now, we just have the
title.  I’ll be working towards getting the thumbnails going later on.
Code is in place to use the Youtube Data API, but is not currently
functional.  Next is to get the Group list for a user so I can then
test getting current queue for a group as well.