Voting code in place
Looks like server is having issues with voting, but the code is in
place.  Can vote initially, but has trouble with un-voting and changing
a vote from one to the other.  Will talk to Josh, but code is in place
for now.  Having weird issues with thumbnails switching around, as well.